feat: remove APIs, types and factories related to DealerPromotion since the product has been removed
BREAKING CHANGE: dealer promotion APIs removed
