+++ src/types/models/dealer.ts
@@ -1,0 +1,11 @@
+export interface DealerLocation {
+  address: string
+  city: string
+  zipCode: string
+}
+
+export interface SearchDealer {
+  id: number
+  location: DealerLocation
+  name: string
+}
